Add WeatherAPI endpoint and use it in the WeatherView
In the end, those views shouldn't be necessary (maybe only for debbuging
purposes) and the API should be used directly in the index page. But, as
an intermediate step, we will switch the module views first.
